requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.7
django-appconf==1.0.6
django-celery-beat==2.5.0
django-channels==0.7.0
django-ckeditor==6.7.0
django-filter==23.5
django-imagekit==5.0.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-resized==1.0.2
django-timezone-field==6.1.0
gunicorn==21.2.0
idna==3.6
kombu==5.3.5
oauthlib==3.2.2
packaging==23.2
pilkit==3.0
pillow==10.2.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
